speaker: move init/deinit to speaker code and backends
