+++ src/base/core/emu.c
@@ -517,20 +517,6 @@ static void __leavedos_main(int code, int sig)
 #endif
     if (config.cpu_vm == CPUVM_KVM || config.cpu_vm_dpmi == CPUVM_KVM)
       kvm_done();
-    if (config.speaker == SPKR_EMULATED) {
-      g_printf("SPEAKER: sound off\n");
-      speaker_off();		/* turn off any sound */
-    }
-    else if (config.speaker==SPKR_NATIVE) {
-       g_printf("SPEAKER: sound off\n");
-       /* Since the speaker is native hardware use port manipulation,
-	* we don't know what is actually implementing the kernel's
-	* ioctls.
-	* My port logic is actually stolen from kd_nosound in the kernel.
-	* 		--EB 21 September 1997
-	*/
-        port_outb(0x61, port_inb(0x61)&0xFC); /* turn off any sound */
-    }
 
     free(vm86_hlt_state);
