+++ src/base/speaker/console_speaker.c
@@ -9,8 +9,9 @@
 #include "Sys/kd.h"
 #endif
 
+#include "emu.h"
 
-void console_speaker_on(void *gp, unsigned ms, unsigned short period)
+static void console_speaker_on(void *gp, unsigned ms, unsigned short period)
 {
 #ifdef __linux__
 	ioctl((int)(uintptr_t)gp, KDMKTONE,
@@ -18,9 +19,16 @@ void console_speaker_on(void *gp, unsigned ms, unsigned short period)
 #endif
 }
 
-void console_speaker_off(void *gp)
+static void console_speaker_off(void *gp)
 {
 #ifdef __linux__
 	ioctl((int)(uintptr_t)gp, KDMKTONE, 0);
 #endif
+}
+
+void console_speaker_init(void)
+{
+	if (config.speaker == SPKR_EMULATED && console_fd != -1)
+		register_speaker((void *)(uintptr_t)console_fd,
+				 console_speaker_on, console_speaker_off);
 }
